% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ah2.R
\name{print.ah2}
\alias{print.ah2}
\title{print.ah2}
\usage{
\method{print}{ah2}(x, digits=3, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{digits}{Integer indicating the number of decimal places.}

\item{...}{Further arguments ignored in this function.}
}
\value{
returns summary output for class 'ah2'
}
\description{
S3 method for class 'ah2'
}
