% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ahreg.R
\name{print.ahreg}
\alias{print.ahreg}
\title{print.ahreg}
\usage{
\method{print}{ahreg}(x, digits=5, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{digits}{Integer indicating the number of decimal places.}

\item{...}{Further arguments ignored in this function.}
}
\value{
returns summary output for class 'ahreg'
}
\description{
S3 method for class 'ahreg'
}
