% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survAH-package.R
\docType{package}
\name{survAH-package}
\alias{survAH-package}
\title{Survival Data Analysis using Average Hazard}
\description{
The survAH package performs two-sample comparisons based on average hazard with survival weight (AHSW) or general censoring-free incidence rate (CFIR) proposed by Uno and Horiguchi (2023) <doi:10.1002/sim.9651>.
}
\references{
Uno H and Horiguchi M. Ratio and difference of average hazard with survival weight: new measures to quantify survival benefit of new therapy. Statistics in Medicine. 2023;1-17. <doi:10.1002/sim.9651>
}
\seealso{
survival survRM2 surv2sampleComp
}
\author{
Hajime Uno, Miki Horiguchi
}
